Add Endpoint class to be parent of Spec2Vec and MS2DeepScore
They have almost all methods in common.

--- don't change below this line ---
MLOPS-320 #time 1h 20m
