--- tests/assets/gnps_2_equal_spectra.mgf
@@ -1,73 +1,0 @@
-BEGIN IONS
-PEPMASS=377.232
-CHARGE=1
-MSLEVEL=2
-SOURCE_INSTRUMENT=LC-ESI-Maxis II HD Q-TOF Bruker
-FILENAME=JnJ_10uL_B7_GB7_01_2592.mzXML
-SEQ=*..*
-IONMODE=positive
-ORGANISM=GNPS-COLLECTIONS-MISC
-NAME=Eicosanoids_ResolvinD1_C22H32O5 M+H
-PI=Dorrestein/Wu
-DATACOLLECTOR=lfnothias
-SMILES=n/a
-INCHI=n/a
-INCHIAUX=N/A
-PUBMED=n/a
-SUBMITUSER=mwang87
-LIBRARYQUALITY=1
-SPECTRUMID=CCMSLIB00001058177
-SCANS=738
-153.792694	151.0
-175.03894	165.0
-241.088303	158.0
-309.44809	286.0
-321.101929	257.0
-375.889862	74.0
-376.199432	90.0
-376.761658	207.0
-377.230042	8316.0
-377.739319	121.0
-377.870697	70.0
-378.237823	1605.0
-378.529327	59.0
-379.616089	52.0
-411.208832	130.0
-END IONS
-
-BEGIN IONS
-PEPMASS=377.232
-CHARGE=1
-MSLEVEL=2
-SOURCE_INSTRUMENT=LC-ESI-Maxis II HD Q-TOF Bruker
-FILENAME=JnJ_10uL_B7_GB7_01_2592.mzXML
-SEQ=*..*
-IONMODE=positive
-ORGANISM=GNPS-COLLECTIONS-MISC
-NAME=Eicosanoids_ResolvinD1_C22H32O5 M+H
-PI=Dorrestein/Wu
-DATACOLLECTOR=lfnothias
-SMILES=n/a
-INCHI=n/a
-INCHIAUX=N/A
-PUBMED=n/a
-SUBMITUSER=mwang87
-LIBRARYQUALITY=1
-SPECTRUMID=CCMSLIB00001058177
-SCANS=738
-153.792694	151.0
-175.03894	165.0
-241.088303	158.0
-309.44809	286.0
-321.101929	257.0
-375.889862	74.0
-376.199432	90.0
-376.761658	207.0
-377.230042	8316.0
-377.739319	121.0
-377.870697	70.0
-378.237823	1605.0
-378.529327	59.0
-379.616089	52.0
-411.208832	130.0
-END IONS
